feat: adds get all gas stations objects
