+++ FuelPump/FuelPump/Domain/Entities/GasStation/GasStation.swift
@@ -46,24 +46,18 @@ struct GasStation: Equatable, Hashable {
             return "default-icon"
         }
     }
-}
 
-struct GetAllGasStation: Equatable {
-    let date: String
-    let gasStations: [GasStation]
-}
-
-extension GetAllGasStation {
-    func mapToRealmObject() -> GetAllGasStationRealm {
-        let model = GetAllGasStationRealm()
-        model._date = date
-        model.gasStations.append(objectsIn: gasStations.map({ $0.mapToRealmObject() }))
-        return model
-    }
-    
-    static func mapFromRealmObject(_ gasStation: GetAllGasStationRealm) -> GetAllGasStation {
-        return GetAllGasStation(date: gasStation._date,
-                                gasStations: gasStation.gasStations.map({ GasStation.mapFromRealmObject($0) }))
+    var prices: [FuelType: Double] {
+        [
+            .dieselA: dieselA,
+            .dieselB: dieselB,
+            .dieselPremium: dieselPremium,
+            .gasoline95_E5: gasoline95_E5,
+            .gasoline95_E10: gasoline95_E10,
+            .gasoline95_E5_Premium: gasoline95_E5_Premium,
+            .gasoline98_E5: gasoline98_E5,
+            .gasoline98_E10: gasoline98_E10
+        ]
     }
 }
